fix ut bug

+++ src/ray/gcs/gcs_client/test/service_based_gcs_client_test.cc
@@ -1291,7 +1291,9 @@ TEST_F(ServiceBasedGcsClientTest, DISABLED_TestGetActorPerf) {
   }
 
   // Get all actors.
-  auto condition = [this, actor_count]() { return GetAllActors().size() == actor_count; };
+  auto condition = [this, actor_count]() {
+    return (int)GetAllActors().size() == actor_count;
+  };
   EXPECT_TRUE(WaitForCondition(condition, timeout_ms_.count()));
 
   int64_t start_time = current_time_ms();
